requirements.txt
Flask
requests
PyJWT
python-dotenv